+++ Package.swift
@@ -1,4 +1,4 @@
-// swift-tools-version:5.5
+// swift-tools-version:6.2
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
@@ -7,8 +7,8 @@ let package = Package(
   name: "SuperwallKit",
   defaultLocalization: "en",
   platforms: [
-    .iOS(.v13),
-    .macOS(.v10_12),
+    .iOS(.v15),
+    .macOS(.v12),
     .watchOS("6.2")
   ],
   products: [
@@ -19,7 +19,7 @@ let package = Package(
     )
   ],
   dependencies: [
-    .package(url: "https://github.com/superwall/superscript-ios-next", .exact("1.0.15"))
+    .package(url: "https://github.com/superwall/superscript-ios-next", exact: "1.0.15")
   ],
   targets: [
     // Targets are the basic building blocks of a package. A target can define a module or a test suite.
@@ -40,5 +40,6 @@ let package = Package(
       name: "SuperwallKitTests",
       dependencies: ["SuperwallKit"]
     )
-  ]
+  ],
+  swiftLanguageModes: [.v5]
 )