+++ Sources/SuperwallKit/Misc/Extensions/Transaction/Transaction+LatestSince.swift
@@ -34,7 +34,6 @@ extension Transaction {
     for productId: String,
     since purchaseDate: Date
   ) async -> VerificationResult<Transaction>? {
-    #if compiler(>=6.1)
     if #available(iOS 18.4, visionOS 2.4, *) {
       var best: VerificationResult<Transaction>?
 
@@ -55,7 +54,6 @@ extension Transaction {
 
       return best
     }
-    #endif
 
     let verificationResult = await Transaction.currentEntitlement(for: productId)