+++ Sources/SuperwallKit/Paywall/View Controller/PaywallViewController.swift
@@ -1527,7 +1527,7 @@ extension PaywallViewController {
 
     if #available(iOS 15.0, *),
       !deviceHelper.isMac {
-      webView.setAllMediaPlaybackSuspended(false)  // ignore-xcode-12
+      webView.setAllMediaPlaybackSuspended(false)
     }
 
     if webView.loadingHandler.didFailToLoad,
@@ -1671,7 +1671,7 @@ extension PaywallViewController {
 
     if #available(iOS 15.0, *),
       !deviceHelper.isMac {
-      webView.setAllMediaPlaybackSuspended(true)  // ignore-xcode-12
+      webView.setAllMediaPlaybackSuspended(true)
     }
 
     resetPresentationPreparations()