+++ Sources/SuperwallKit/StoreKit/Products/StoreProduct/SK2StoreProduct.swift
@@ -35,11 +35,7 @@ struct SK2StoreProduct: StoreProductType {
     entitlements: Set<Entitlement>,
     billingPlanType: AppStoreProduct.BillingPlanType? = nil
   ) {
-    #if swift(<5.7)
-    self._underlyingSK2Product = sk2Product
-    #else
     self.underlyingSK2Product = sk2Product
-    #endif
     self.entitlements = entitlements
     self.billingPlanType = billingPlanType
 
@@ -174,18 +170,7 @@ struct SK2StoreProduct: StoreProductType {
     )
   }
 
-  #if swift(<5.7)
-  // We can't directly store instances of StoreKit.Product, since that causes
-  // linking issues in iOS < 15, even with @available checks correctly in place.
-  // So instead, we store the underlying product as Any and wrap it with casting.
-  private let _underlyingSK2Product: Any
-  var underlyingSK2Product: SK2Product {
-    // swiftlint:disable:next force_cast
-    _underlyingSK2Product as! SK2Product
-  }
-  #else
   let underlyingSK2Product: SK2Product
-  #endif
 
   var productIdentifier: String {
     underlyingSK2Product.id